Add Super MaD game mode
"Super MaD" was another game I was developing back in the days using
the same engine. It never finished and obviously has some teenager
humor that hasn't aged well. Data file is original but converted to
the new format and other new features disabled but sprint
(because it makes the game so much more fun to play).